[FIX] Replace Double with BigDecimal for monetary values in models and DTOs
- BankAccountEntity: balance field changed from Double to BigDecimal
- CurrencyExchangeRateEntity: exchangeRate field changed from Double to BigDecimal
- BankAccountResponse: balance field updated to BigDecimal
- ExchangeRateResponse: exchangeRate field updated to BigDecimal
- SetExchangeRateRequest: exchangeRate field updated to BigDecimal
- Added @Version field to BankAccountEntity for optimistic locking

+++ src/main/java/ro/unibuc/prodeng/model/CurrencyExchangeRateEntity.java
@@ -3,11 +3,13 @@ package ro.unibuc.prodeng.model;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Document(collection = "currency_exchange_rates")
 public record CurrencyExchangeRateEntity(
     @Id
     String id,
     String sourceCurrency,
     String targetCurrency,
-    Double exchangeRate
+    BigDecimal exchangeRate
 ) {}
