+++ src/main/java/ro/unibuc/prodeng/response/ExchangeRateResponse.java
@@ -1,8 +1,10 @@
 package ro.unibuc.prodeng.response;
 
+import java.math.BigDecimal;
+
 public record ExchangeRateResponse(
     String id,
     String sourceCurrency,
     String targetCurrency,
-    Double exchangeRate
+    BigDecimal exchangeRate
 ) {}
